pyproject.toml
[project]
name = "python-ai-pygame"
version = "0.1.0"
description = "A Python game using pygame-ce"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
